requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
APScheduler==3.10.4
google-generativeai==0.3.2
google-genai==0.2.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
requests-oauthlib==1.3.1
tweepy==4.14.0
facebook-sdk==3.1.0
python-linkedin-v2==0.9.0